+++ yolox/utils/boxes.py
@@ -100,7 +100,7 @@ def bboxes_iou(bboxes_a, bboxes_b, xyxy=True):
         area_a = torch.prod(bboxes_a[:, 2:], 1)
         area_b = torch.prod(bboxes_b[:, 2:], 1)
     device, dtype = parse_dtype(tl.type())
-    en = (tl < br).type(dtype).to(device=device).prod(dim=2)
+    en = (tl < br).to(device=device, dtype=dtype).prod(dim=2)
     area_i = torch.prod(br - tl, 2) * en  # * ((tl < br).all())
     return area_i / (area_a[:, None] + area_b - area_i)
